Reconstruction of Under-sampled Super High Frequency Signals using Recurrent Nonuniform Sampling
Sampling devices form an interface between front ends of radio receivers and signal processing units and therefore play a vital role in digital radio and radar communications. In order to obtain signal outputs from devices without any loss of information, effective signal reconstruction is essential. In general, a signal can be faithfully reconstructed if it is sampled at a rate greater than or equal to the Nyquist rate. In some high frequency applications however, it is not possible to sample the signal at the Nyquist rate. In this thesis, we deal with the under-sampling and effective reconstruction of very high frequency signals. Here, we first further the discussion of accurately determining the center frequency using the main lobe to side lobe ratio, as discussed in paper [1], and then develop a practical implementation for the same. Next, we tackle the problem of aliasing, faced when nonuniformly sampling the signal at one-third and two-third of the sampling frequency. The aim is to develop a practical solution that comes within 3% of the theoretical results. Finally, we reconstruct the input signal with the help of bandpass filtering (to remove the side lobes) and interpolation.